feat: initial arcade ui component refactor
