+++ lib/saito/new-ui/modals/user-menu/user-menu.js
@@ -17,9 +17,9 @@ class UserMenu {
     render(app) {
         if (!document.querySelector("#saito-user-menu")) {
             this.overlay.show(app, "", userMenuTemplate(app));
-            let mods = app.modules.respondTo("user_menu");
+            let mods = app.modules.respondTo("user-menu");
             mods.forEach((mod, index, arrry) => {
-                let item = mod.respondTo('user_menu');
+                let item = mod.respondTo('user-menu');
                 let id = `user_menu_item_${index}`
                 this.callbacks[id] = item.callback
                 this.addMenuItem(item, id);
@@ -43,9 +43,8 @@ class UserMenu {
 
     addMenuItem(item, id) {
         document.querySelector(".saito-modal-content").innerHTML += `
-        <div id="${id}" class="user_menu_item saito-modal-option"><i class="${item.icon}"></i><div>${item.text}</div></div>
-     
-        `
+          <div id="${id}" class="user_menu_item saito-modal-option"><i class="${item.icon}"></i><div>${item.text}</div></div>
+        `;
     }
 }
