+++ mods/arcade/lib/game/main.template.js
