+++ mods/arcade/lib/main/main.js
@@ -1,11 +1,12 @@
-const ArcadeMainTemplate = require("./templates/arcade-main.template");
-const ArcadeContainerTemplate = require("./templates/arcade-container.template");
-const ArcadeMobileHelper = require("./templates/arcade-mobile-helper.template");
-const ArcadeForums = require("./arcade-forums");
-const ArcadePosts = require("./arcade-posts");
+const ArcadeMainTemplate = require("./main.template");
+
+const ArcadeContainerTemplate = require("./../templates/arcade-container.template");
+const ArcadeMobileHelper = require("./../templates/arcade-mobile-helper.template");
+const ArcadeForums = require("./forums");
+const ArcadePosts = require("./posts");
 const GameLoader = require("./../../../../lib/saito/new-ui/game-loader/game-loader");
 const SaitoCarousel = require("./../../../../lib/saito/ui/saito-carousel/saito-carousel");
-const ArcadeInviteTemplate = require("./templates/arcade-invite.template");
+const ArcadeInviteTemplate = require("./../templates/arcade-invite.template");
 const GameCryptoTransferManager = require("./../../../../lib/saito/new-ui/game-crypto-transfer-manager/game-crypto-transfer-manager");
 const JSON = require("json-bigint");
 const saito = require("../../../../lib/saito/saito");
