+++ mods/arcade/lib/main/main.template.js
