+++ mods/arcade/lib/sidebar/arcade-game-sidebar.template.js
