+++ mods/arcade/lib/sidebar/main.template.js
