+++ mods/arcade/lib/templates/arcade-challenge.template.js
