+++ mods/arcade/lib/templates/arcade-container.template.js
