+++ mods/arcade/lib/templates/arcade-forums-thread.template.js
