+++ mods/arcade/lib/templates/arcade-forums.template.js
