+++ mods/arcade/lib/templates/arcade-invite.template.js
