+++ mods/arcade/lib/templates/arcade-mobile-helper.template.js
