+++ mods/arcade/lib/templates/arcade-posts.template.js
