+++ mods/arcade/lib/templates/game-create-menu.template.js
