+++ mods/redsquare/redsquare.js
@@ -106,7 +106,7 @@ class RedSquare extends ModTemplate {
       };
     }
 
-    if (type == "user_menu") {
+    if (type == "user-menu") {
       return {
         text: "Add Contact",
         icon: "far fa-id-card",
